+++ packages/faucet/src/constants.ts
@@ -4,8 +4,7 @@ export const binaryName = "cosmwasm-faucet";
 export const concurrency: number = Number.parseInt(process.env.FAUCET_CONCURRENCY || "", 10) || 5;
 export const port: number = Number.parseInt(process.env.FAUCET_PORT || "", 10) || 8000;
 export const mnemonic: string | undefined = process.env.FAUCET_MNEMONIC;
-
-export const addressPrefix = "cosmos";
+export const addressPrefix = process.env.FAUCET_ADDRESS_PREFIX || "cosmos";
 
 /** For the local development chain */
 export const developmentTokenConfig: TokenConfiguration = {